requirements.txt
Django
# djangorestframework
# django-seed
django-crispy-forms
django-ckeditor
pillow
django-allauth
django-notifications-hq
# django-model-utils
six
django-preventconcurrentlogins
